Replace Netty with OkHttp
Netty is a NIO client server framework which depends on the system
specific native implementations of the JVM. In my experience this
results in different behavior on different platforms. OkHttp uses byte
arrays internal with a clear API. Its TLS implementation is very well
integrated. LittleProxy and LittleProxy-mitm dependencies are obsolete.
This is a drop in replacement of the http proxy in MoCuishle based on
OkHttp. It's necessary to import the new generated certificate agency
~/MoCuishle/okproxy.pem into the browser.