Replace Netty with OkHttp (#10)
Netty is a NIO client server framework which depends on the system
specific native implementations of the JVM. In my experience this
results in different behavior on different platforms. OkHttp uses byte
arrays internal with a clear API. Its TLS implementation is very well
integrated. LittleProxy and LittleProxy-mitm dependencies are obsolete.
This is a drop in replacement of the http proxy in MoCuishle based on
OkHttp.

- Reviewed  Maven build, updated dependencies, renamed modules
- Fixes #5 built with current Java and Maven, but Java 11 too
- Integrate JaCoCo coverage report in Maven build
- Use maven.compiler.release to fix warnings
- Reuse mocuishle KeyStore and CA, but allow configuration by properties
- Reviewed README.md files